Move NVTX profiling so it can run in the shared provider properly
