+++ onnxruntime/core/providers/cuda/cuda_provider_factory.h
@@ -15,6 +15,9 @@ struct CUDAExecutionProviderExternalAllocatorInfo;
 namespace cuda {
 class INcclService;
 }
+namespace profile {
+  class NvtxRangeCreator;
+}
 
 struct ProviderInfo_CUDA {
   virtual OrtStatus* SetCurrentGpuDeviceId(_In_ int device_id) = 0;
@@ -40,6 +43,11 @@ struct ProviderInfo_CUDA {
 
 #if defined(USE_CUDA) && defined(ORT_USE_NCCL) && defined(USE_NCCL_P2P)
   virtual onnxruntime::cuda::INcclService& GetINcclService() = 0;
+#endif
+
+#ifdef ENABLE_NVTX_PROFILE
+  virtual void NvtxRangeCreator__BeginImpl(profile::NvtxRangeCreator *p) = 0;
+  virtual void NvtxRangeCreator__EndImpl(profile::NvtxRangeCreator* p) = 0;
 #endif
 
   virtual std::shared_ptr<onnxruntime::IExecutionProviderFactory> CreateExecutionProviderFactory(const onnxruntime::CUDAExecutionProviderInfo& info) = 0;
